Move IApiServiceFactory to IdeaStatica.Api
