+++ src/api-sdks/connection-api/clients/csharp/src/IdeaStatiCa.ConnectionApi/ConnectionApiServiceRunner.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using IdeaStatiCa.Api.Common;
+using System;
 using System.Diagnostics;
 using System.IO;
 using System.Net;
